product 엔티티, 레파지토리 기본키타입 int 로 변경

+++ backend/src/main/java/com/project/backend/repository/PaymentRepository.java
@@ -4,6 +4,6 @@ import com.project.backend.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
 }
